NG-RO/requirements.txt
##
# Licensed under the Apache License, Version 2.0 (the "License"); you may
# not use this file except in compliance with the License. You may obtain
# a copy of the License at
#
#         http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
# WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
# License for the specific language governing permissions and limitations
# under the License.
##

PyYAML
CherryPy==18.1.2
osm-common @ git+https://osm.etsi.org/gerrit/osm/common.git#egg=osm-common
requests
cryptography
osm-im @ git+https://osm.etsi.org/gerrit/osm/IM.git#egg=osm-im
osm-ro-plugin @ git+https://osm.etsi.org/gerrit/osm/RO.git#egg=osm-ro-plugin&subdirectory=RO-plugin
logutils
